requirements.txt
python-pptx==0.6.21
pyyaml==6.0.1
openai>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
